LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-error-2-expected.txt

CONSOLE MESSAGE: ReferenceError: Cannot access uninitialized variable.

FAIL Test that ill-founded cyclic dependencies cause ReferenceError during evaluation, which leads to error events on window, and that exceptions are remembered.
 assert_array_equals: lengths differ, expected array ["cycle-tdz-access-a", object "ReferenceError: Cannot access uninitialized variable.", [...], [...]] length 4, got ["cycle-tdz-access-a", object "ReferenceError: Cannot access uninitialized variable."] length 2
